fix(fe): resolve frontend lint errors
