+++ HSTS.FE/src/routes/RouteShell.jsx
@@ -1,0 +1,16 @@
+import React, { Suspense } from 'react';
+import { Spin } from 'antd';
+
+const LoadingFallback = () => (
+  <div style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100vh' }}>
+    <Spin size="large" tip="Loading page..." />
+  </div>
+);
+
+const SuspenseWrapper = ({ children }) => (
+  <Suspense fallback={<LoadingFallback />}>
+    {children}
+  </Suspense>
+);
+
+export default SuspenseWrapper;
